# flex/coordinator/gs_flex_coordinator/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: graphscope@alibaba-inc.com
    name: GraphScope
  description: |-
    This is a specification for GraphScope FLEX HTTP service based on the OpenAPI 3.0 specification. You can find out more details about
    specification at [doc](https://swagger.io/specification/v3/).

    Some useful links:
    - [GraphScope Repository](https://github.com/alibaba/GraphScope)
    - [The Source API definition for GraphScope Interactive](https://github.com/GraphScope/portal/tree/main/httpservice)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GraphScope FLEX HTTP SERVICE API
  version: 0.9.1
...
servers:
- url: /
paths:
  /api/v1/alert/message:
    get:
      operationId: list_alert_messages
      parameters:
      - explode: true
        in: query
        name: alert_type
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - solved
          - unsolved
          - dealing
          type: string
        style: form
      - explode: true
        in: query
        name: severity
        required: false
        schema:
          enum:
          - warning
          - emergency
          type: string
        style: form
      - description: format with "2023-02-21-11-56-30"
        explode: true
        in: query
        name: start_time
        required: false
        schema:
          type: string
        style: form
      - description: format with "2023-02-21-11-56-30"
        explode: true
        in: query
        name: end_time
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - message_id: ALERT-MESSAGE-QWERTYUIOP-1
                alert_name: HighDiskUtilization
                severity: warning
                metric_type: node
                target:
                - hostname-1
                trigger_time: 2023-02-21 11:56:30
                status: unsolved
                message: "Disk usage node1 85.12%, exceeds threshold"
              - message_id: ALERT-MESSAGE-QWERTYUIOP-2
                alert_name: HighDiskUtilization
                severity: emergency
                metric_type: node
                target:
                - hostname-1
                trigger_time: 2023-02-21 11:56:30
                status: solved
                message: "Disk usage node1 85.12%, exceeds threshold"
              - message_id: ALERT-MESSAGE-QWERTYUIOP-3
                alert_name: HighDiskUtilization
                severity: warning
                metric_type: node
                target:
                - hostname-1
                trigger_time: 2023-02-21 11:56:30
                status: dealing
                message: "Disk usage node1 85.12%, exceeds threshold"
              schema:
                items:
                  $ref: '#/components/schemas/AlertMessage'
                type: array
          description: Successful operation
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
    put:
      description: Update alert messages in batch
      operationId: update_alert_messages
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/update_alert_messages_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
        "404":
          description: Alert message not found
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
  /api/v1/alert/receiver:
    get:
      operationId: list_receivers
      responses:
        "200":
          content:
            application/json:
              example:
              - type: webhook
                receiver_id: RECEIVER-QWERTYUIOP-1
                webhook_url: http://webhook.com
                is_at_all: false
                at_user_ids:
                - 111111
                - 222222
                enable: false
              - type: webhook
                receiver_id: RECEIVER-QWERTYUIOP-2
                webhook_url: http://webhook.com
                is_at_all: false
                at_user_ids:
                - 111111
                - 222222
                enable: true
                message: error message
              schema:
                items:
                  $ref: '#/components/schemas/AlertReceiver'
                type: array
          description: Successful operation
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
    post:
      operationId: register_receiver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertReceiver'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
  /api/v1/alert/receiver/{receiver_id}:
    delete:
      operationId: delete_receiver_by_id
      parameters:
      - explode: false
        in: path
        name: receiver_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
        "404":
          description: Receiver not found
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
    put:
      operationId: update_receiver_by_id
      parameters:
      - explode: false
        in: path
        name: receiver_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertReceiver'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
        "404":
          description: Receiver not found
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
...
  /api/v1/alert/rule/{rule_name}:
    delete:
      operationId: delete_alert_rule_by_name
      parameters:
      - explode: false
        in: path
        name: rule_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
    put:
      operationId: update_alert_rule_by_name
      parameters:
      - explode: false
        in: path
        name: rule_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - alert
      x-openapi-router-controller: gs_flex_coordinator.controllers.alert_controller
...
components:
  schemas:
    ApiResponse:
      example: operation successful
      type: string
...
    AlertRule:
      example:
        severity: warning
        conditions_description: conditions_description
        enable: true
        name: name
        metric_type: node
        frequency: 0
      properties:
        name:
          title: name
          type: string
        severity:
          enum:
          - warning
          - emergency
          title: severity
          type: string
        metric_type:
          enum:
          - node
          - service
          title: metric_type
          type: string
        conditions_description:
          title: conditions_description
          type: string
        frequency:
          description: (mins)
          title: frequency
          type: integer
        enable:
          title: enable
          type: boolean
      title: AlertRule
      type: object
      x-body-name: alert_rule
    AlertMessage:
      example:
        severity: warning
        trigger_time: trigger_time
        metric_type: node
        message_id: message_id
        message: message
        alert_name: alert_name
        target:
        - target
        - target
        status: unsolved
      properties:
        message_id:
          description: Generated in server side
          title: message_id
          type: string
        alert_name:
          title: alert_name
          type: string
        severity:
          enum:
          - warning
          - emergency
          title: severity
          type: string
        metric_type:
          enum:
          - node
          - service
          title: metric_type
          type: string
        target:
          items:
            type: string
          title: target
          type: array
        trigger_time:
          title: trigger_time
          type: string
        status:
          enum:
          - unsolved
          - solved
          - dealing
          title: status
          type: string
        message:
          title: message
          type: string
      title: AlertMessage
      type: object
      x-body-name: alert_message
    AlertReceiver:
      example:
        webhook_url: webhook_url
        is_at_all: true
        receiver_id: receiver_id
        enable: true
        at_user_ids:
        - at_user_ids
        - at_user_ids
        type: webhook
        message: message
      properties:
        type:
          enum:
          - webhook
          title: type
          type: string
        receiver_id:
          description: Generated in server side
          title: receiver_id
          type: string
        webhook_url:
          title: webhook_url
          type: string
        at_user_ids:
          items:
            type: string
          title: at_user_ids
          type: array
        is_at_all:
          title: is_at_all
          type: boolean
        enable:
          title: enable
          type: boolean
        message:
          description: error message generated in server side
          title: message
          type: string
      title: AlertReceiver
      type: object
      x-body-name: alert_receiver
...
    update_alert_messages_request:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/AlertMessage'
          title: messages
          type: array
        batch_status:
          description: Override the status of each message
          enum:
          - solved
          - unsolved
          - dealing
          title: batch_status
          type: string
        batch_delete:
          default: false
          description: True will delete all the messages in request body
          title: batch_delete
          type: boolean
      title: update_alert_messages_request
      type: object
...